implemented single point of navigation, refactored entity store to use reducer hook
