+++ ui/app/lib/entity-store/entity-actions.ts
@@ -1,4 +1,4 @@
-import { Entity, EntityStore } from './entity-store';
+import { Entity, EntityStore } from "./entity-store";
 
 export type EntityActions<T> = {
   addEntity: (entity: Entity<T>) => void;
@@ -17,36 +17,40 @@ export const createActions = <T>(
   state: EntityStore<T>,
   setState: (state: EntityStore<T>) => void
 ): EntityActions<T> => ({
-    addEntity: (entity: Entity<T>) => {
-      console.log('setting', entity, state)
-      setState({
-        ...state,
-        entities: { ...state.entities, [entity.id]: entity }
-      })
-    }
-    ,
-    addEntities: (entities: { [entityId: string]: Entity<T> }) =>
-      setState({ ...state, entities: { ...state.entities, ...entities } }),
-    removeEntity: (entityId: string) => {
-      const { [entityId]: _, ...rest } = state.entities 
-      setState({
-        ...state,
-        entities: { ...rest }
-      })
-    },
-    clearEntities: () =>
-      setState({
-        ...state,
-        entities: {},
-        selectedEntityId: null
-      }),
-    setLoading: (isLoading = true) => setState({ ...state, isLoading }),
-    setError: (error: string) => setState({ ...state, error }),
-    setStatus: (status: string) => console.log('disabled'),
-    setSelectedEntityId: (selectedEntityId: string) =>
-      setState({ ...state, selectedEntityId }),
-    selectEntity: (entity: Entity<T>) =>
-      setState({ ...state, selectedEntityId: entity.id }),
-    getEntity: (entityId: string): Entity<T> | undefined =>
-      state.entities[entityId]
-  });
+  addEntity: (entity: Entity<T>) => {
+    console.log("setting", entity, state);
+    setState({
+      type: "add",
+      payload: {
+        entities: {
+          [entity.id]: entity,
+        },
+      },
+    });
+  },
+  addEntities: (entities: { [entityId: string]: Entity<T> }) =>
+    setState({ ...state, entities: { ...state.entities, ...entities } }),
+  removeEntity: (entityId: string) => {
+    const { [entityId]: _, ...rest } = state.entities;
+    setState({
+      ...state,
+      entities: { ...rest },
+    });
+  },
+  clearEntities: () =>
+    setState({
+      ...state,
+      entities: {},
+      selectedEntityId: null,
+    }),
+  setLoading: (isLoading = true) => setState({ ...state, isLoading }),
+  setError: (error: string) => setState({ ...state, error }),
+  setStatus: (status: string) =>
+    setState({ type: "add", payload: { status: status } }),
+  setSelectedEntityId: (selectedEntityId: string) =>
+    setState({ ...state, selectedEntityId }),
+  selectEntity: (entity: Entity<T>) =>
+    setState({ ...state, selectedEntityId: entity.id }),
+  getEntity: (entityId: string): Entity<T> | undefined =>
+    state.entities[entityId],
+});
