+++ ui/app/pages/flags/flags.tsx
@@ -1,20 +1,14 @@
+import Layout from 'antd/lib/layout/layout';
 import React from 'react';
 
 import AppNavigation from '../../../components/app-navigation';
-import PageLayout from '../../../components/page-layout';
 
 const Flags: React.FC = () => {
   return (
-    <PageLayout
-      navigation={
-        <AppNavigation
-          title="Flags"
-          hasBackIcon
-        />
-      }
+    <Layout
     >
       Todo: List of flags
-    </PageLayout>
+    </Layout>
   );
 };
