+++ ui/app/pages/instances/instances.style.tsx
@@ -1,0 +1,8 @@
+import { Button } from "antd";
+import styled from "@emotion/styled";
+
+export const SmallButton = styled(Button)`
+  display: inline-block;
+  width: fit-content;
+  margin-bottom: 15px;
+`;
