+++ ui/app/pages/projects/projects.tsx
@@ -1,20 +1,12 @@
+import Layout from 'antd/lib/layout/layout';
 import React from 'react';
-
-import AppNavigation from '../../../components/app-navigation';
-import PageLayout from '../../../components/page-layout';
 
 const Projects: React.FC = () => {
   return (
-    <PageLayout
-      navigation={
-        <AppNavigation
-          title="Projects"
-          hasBackIcon
-        />
-      }
+    <Layout
     >
       Todo: List of projects
-    </PageLayout>
+    </Layout>
   );
 };
