+++ ui/app/pages/segments/segments.tsx
@@ -1,20 +1,12 @@
+import { Layout } from 'antd';
 import React from 'react';
-
-import AppNavigation from '../../../components/app-navigation';
-import PageLayout from '../../../components/page-layout';
 
 const Segments: React.FC = () => {
   return (
-    <PageLayout
-      navigation={
-        <AppNavigation
-          title="Segments"
-          hasBackIcon
-        />
-      }
+    <Layout
     >
       Todo: List of segments
-    </PageLayout>
+    </Layout>
   );
 };
