+++ ui/components/app-navigation/app-navigation.types.tsx
@@ -1,0 +1,29 @@
+import { PageHeaderProps } from "antd";
+import { ReactElement } from "react";
+
+export interface ButtonProps {
+  title: string;
+  type: string;
+}
+
+export type FlagbaseSubMenuProps =
+  | "Home"
+  | "Instance"
+  | "Workspace"
+  | "Project"
+  | "Flags";
+
+export type FlagbaseSubMenuValues = {
+  title: ReactElement;
+  redirect: string;
+  content: { title: string; href: string }[];
+};
+
+export type AppNavigationProps = {
+  title: string;
+  hasBackIcon?: boolean;
+} & PageHeaderProps;
+
+export type SubMenuProps = {
+  subMenuContent: Partial<Record<keyof FlagbaseSubMenuProps, string>>;
+};
